restructure project in mapviewstory package #23
